Modularize salvage and externalize mining tuning
