+++ lib/mining.cfg
@@ -1,0 +1,21 @@
+# Mining tuning. Values reproduce the legacy placement and reward behavior.
+mining.region.map.start=500000
+mining.region.map.end=659999
+mining.region.map.duration=9
+mining.region.ud.start=700000
+mining.region.ud.end=859999
+mining.region.ud.duration=9
+mining.region.tharnrift.start=110000
+mining.region.tharnrift.end=119999
+mining.region.tharnrift.duration=9
+mining.region.mapg.start=500000
+mining.region.mapg.end=659999
+mining.region.mapg.duration=15
+mining.region.udg.start=700000
+mining.region.udg.end=859999
+mining.region.udg.duration=15
+# Optional per-region capacity override: mining.region.map.max=12
+# Weighted gem pool; all omitted entries retain their legacy weights.
+mining.gem.504.weight=8
+mining.gem.505.weight=8
+mining.gem.506.weight=6
